Knight Changes
Took a break from looking at ik for now and kinda went crazy with the knight script; added a fuck ton to that thing but still need to figure out how to make the knight's burning system to work the way i want it to. and need to ensure that when a player dies i dont disable their scripts just the hitbox camera and mesh.