Worked on knight smoke
Was working on it finished most of the code then realized i didnt really like the direction i was taking it so im stopping to think of a better way of going about it. rn it works with the knight throwing out a  smoke grenade type thing and it explodes with a firey smoke left behind but rn i think it would feel more interactive if it comes out of his hand and if the player moves the mouse the smoke follows the path for a second similar to that one guy in valorant but we will see what i go for.